ps_text_remover/requirements.txt
pyautogui==0.9.54
pillow==10.0.0
opencv-python==4.8.0.74
pyinstaller==6.0.0
